+++ Modules/Office365DSC/DSCResources/MSFT_SPOSearchManagedProperty/MSFT_SPOSearchManagedProperty.psm1
@@ -89,18 +89,13 @@ function Get-TargetResource
         $Ensure = "Present",
 
         [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
-
-        [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
         $GlobalAdminAccount
     )
 
     Write-Verbose -Message "Getting configuration for Managed Property instance $Name"
 
-    Test-MSCloudLogin -ConnectionUrl $CentralAdminUrl `
-                      -O365Credential $GlobalAdminAccount `
+    Test-MSCloudLogin -O365Credential $GlobalAdminAccount `
                       -Platform SharePointOnline
 
     $nullReturn = @{
@@ -124,7 +119,6 @@ function Get-TargetResource
         MappedCrawledProperties     = $null
         CompanyNameExtraction       = $null
         Ensure                      = "Absent"
-        CentralAdminUrl             = $CentralAdminUrl
     }
 
     if ($null -eq $Script:RecentMPExtract)
@@ -194,7 +188,6 @@ function Get-TargetResource
         FinerQueryTokenization = [boolean] $property.Value.ExpandSegments
         MappedCrawledProperties = $mappings
         CompanyNameExtraction = $CompanyNameExtraction
-        CentralAdminUrl = $CentralAdminUrl
         Ensure = "Present"
     }
 }
@@ -289,18 +282,13 @@ function Set-TargetResource
         $Ensure = "Present",
 
         [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
-
-        [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
         $GlobalAdminAccount
     )
 
     Write-Verbose -Message "Setting configuration for Managed Property instance $Name"
 
-    Test-MSCloudLogin -ConnectionUrl $CentralAdminUrl `
-                      -O365Credential $GlobalAdminAccount `
+    Test-MSCloudLogin -O365Credential $GlobalAdminAccount `
                       -Platform SharePointOnline
 
     $SearchConfigTemplatePath =  Join-Path -Path $PSScriptRoot `
@@ -705,10 +693,6 @@ function Test-TargetResource
         $Ensure = "Present",
 
         [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
-
-        [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
         $GlobalAdminAccount
     )
@@ -745,10 +729,6 @@ function Export-TargetResource
         [ValidateSet("Text","Integer","Decimal","DateTime","YesNo","Double","Binary")]
         [System.String]
         $Type,
-
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]