+++ Modules/Office365DSC/DSCResources/MSFT_SPOTheme/MSFT_SPOTheme.psm1
@@ -22,18 +22,13 @@ function Get-TargetResource
         $Ensure = "Present",
 
         [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
-
-        [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
         $GlobalAdminAccount
     )
 
     Write-Verbose -Message "Getting configuration for SPO Theme $Name"
 
-    Test-MSCloudLogin -ConnectionUrl $CentralAdminUrl `
-                      -O365Credential $GlobalAdminAccount `
+    Test-MSCloudLogin -O365Credential $GlobalAdminAccount `
                       -Platform SharePointOnline
 
     $nullReturn = @{
@@ -41,7 +36,6 @@ function Get-TargetResource
         IsInverted          = $null
         Palette             = $null
         Ensure              = "Absent"
-        CentralAdminUrl     = $CentralAdminUrl
         GlobalAdminAccount  = $GlobalAdminAccount
     }
 
@@ -59,7 +53,6 @@ function Get-TargetResource
             Name                = $theme.Name
             IsInverted          = $theme.IsInverted
             Palette             = $theme.Palette
-            CentralAdminUrl     = $CentralAdminUrl
             GlobalAdminAccount  = $GlobalAdminAccount
             Ensure              = "Present"
         }
@@ -94,18 +87,13 @@ function Set-TargetResource
         $Ensure = "Present",
 
         [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
-
-        [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
         $GlobalAdminAccount
     )
 
     Write-Verbose -Message "Setting configuration for SPO Theme $Name"
 
-    Test-MSCloudLogin -ConnectionUrl $CentralAdminUrl `
-                      -O365Credential $GlobalAdminAccount `
+    Test-MSCloudLogin -O365Credential $GlobalAdminAccount `
                       -Platform SharePointOnline
 
     if ($Ensure -eq "Present")
@@ -119,7 +107,6 @@ function Set-TargetResource
         $CurrentParameters = $PSBoundParameters
         $CurrentParameters.Remove("Palette")
         $CurrentParameters.Remove("Ensure")
-        $CurrentParameters.Remove("CentralAdminURL")
         $CurrentParameters.Remove("GlobalAdminAccount")
         $CurrentParameters.Add("Palette", $PaletteHash)
 
@@ -194,10 +181,6 @@ function Test-TargetResource
         $Ensure = "Present",
 
         [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
-
-        [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
         $GlobalAdminAccount
     )
@@ -229,10 +212,6 @@ function Export-TargetResource
         [Parameter(Mandatory = $true)]
         [System.String]
         $Name,
-
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $CentralAdminUrl,
 
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]