+++ Modules/Office365DSC/Examples/Resources/SPOSearchResultSource/1-CreateNewResultSource.ps1
@@ -22,7 +22,6 @@ Configuration Example
             Type               = "SharePoint"
             GlobalAdminAccount = $credsGlobalAdmin
             Ensure             = "Present"
-            CentralAdminUrl    = "https://Office365DSC-admin.sharepoint.com"
         }
     }
 }