+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 ## Unreleased
 
+* BREAKING CHANGES
+  * ODSettings
+    * IsSingleInstance is now a Mandatory parameter;
+    * Removed CentralAdminUrl as a parameter;
+  * SPOAccessControlSettings
+    * Removed CentralAdminUrl as a parameter;
+  * SPOAPP
+    * Removed CentralAdminUrl as a parameter;
+  * SPOHubSite
+    * Removed CentralAdminUrl as a parameter;
+  * SPOSearchManagedProperty
+    * Removed CentralAdminUrl as a parameter;
+  * SPOSearchResultSource
+    * Removed CentralAdminUrl as a parameter;
+  * SPOSharingSettings
+    * Removed CentralAdminUrl as a parameter;
+  * SPOSite
+    * Removed CentralAdminUrl as a parameter;
+  * SPOSiteDesign
+    * Removed CentralAdminUrl as a parameter;
+  * SPOSiteDesignRights
+    * Removed CentralAdminUrl as a parameter;
+  * SPOTenantSettings
+    * Removed CentralAdminUrl as a parameter;
+  * SPOTheme
+    * Removed CentralAdminUrl as a parameter;
+* SPOApp
+  * Change logic for detection when no App Catalog exist;
+
+## 1.0.0.744
+
 * GENERAL
   * Updated Dependency on SharePointPnPPowerShellOnline
     to version 3.11.1907.0